0.8.0
The probability of having a mutation is now part of the parameters.
Can be turned to 1 with a small SD to get constant but diffuse mutation, or small with high SD (as in the original paper).

Fixes #19, btw :)
We now have branches, woo-hoo!
The issue was that mutations were way too frequent, causing a high ""gene flow"" that prevented branching.